Renamed AppConfig namespace to ApplicationConfig
