+++ SMB3Explorer/ViewModels/LandingViewModel.cs
@@ -8,8 +8,8 @@ using CommunityToolkit.Mvvm.Input;
 using OneOf;
 using OneOf.Types;
 using Serilog;
-using SMB3Explorer.AppConfig;
-using SMB3Explorer.AppConfig.Models;
+using SMB3Explorer.ApplicationConfig;
+using SMB3Explorer.ApplicationConfig.Models;
 using SMB3Explorer.Enums;
 using SMB3Explorer.Models.Internal;
 using SMB3Explorer.Services.ApplicationContext;
